# Conflicts:
#	Hotline/HotlineConfigStrings.h
#	HotlineSDK/HotlineSDK/Hotline.m
#	HotlineSDK/HotlineSDK/Network/DataUpdater/FDMessagesUpdater.m
#	HotlineSDK/HotlineSDK/Network/DataUpdater/FDSessionUpdater.m
#	HotlineSDK/HotlineSDK/Network/Services/HLCoreServices.m
#	HotlineSDK/HotlineSDK/Network/Services/HLMessageServices.m
#	HotlineSDK/HotlineSDK/Utilities/FDUtilities.m

+++ HotlineSDK/HotlineSDK/Network/Services/HLMessageServices.m
@@ -31,6 +31,9 @@
 #import "Message.h"
 #import "Fragment.h"
 #import "FCRemoteConfigUtil.h"
+#import "FDLocaleUtil.h"
+#import "FDConstants.h"
+#import "HLUserDefaults.h"
 #import "HLUser.h"
 
 #define ERROR_CODE_USER_NOT_CREATED -1
@@ -311,23 +314,38 @@ static HLNotificationHandler *handleUpdateNotification;
     NSString *path = [NSString stringWithFormat:HOTLINE_API_CHANNELS_PATH,appID];
     NSString *token = [NSString stringWithFormat:HOTLINE_REQUEST_PARAMS,appKey];
     NSNumber *lastUpdateTime = [FDUtilities getLastUpdatedTimeForKey:HOTLINE_DEFAULTS_CHANNELS_LAST_UPDATED_SERVER_TIME];
-    NSString *afterTime = [NSString stringWithFormat:@"after=%@",lastUpdateTime];
-    
-    [request setRelativePath:path andURLParams:@[token, @"tags=true", afterTime]];
+    NSString *afterTime = [NSString stringWithFormat:PARAM_SINCE,lastUpdateTime];
+    NSNumber *requestlocaleId = [FDLocaleUtil getConvLocaleId];
+    NSMutableArray *reqParams = [[NSMutableArray alloc]initWithArray:@[token,afterTime]];
+    [reqParams addObjectsFromArray:[FDLocaleUtil channelLocaleParams]];
+    [request setRelativePath:path andURLParams:reqParams];
     NSURLSessionDataTask *task = [apiClient request:request withHandler:^(FDResponseInfo *responseInfo, NSError *error) {
-        if (!error) {
+        NSInteger statusCode = ((NSHTTPURLResponse *)responseInfo.response).statusCode;
+        if (!error && statusCode == 200) {
             /* This check is added to delete all messages that are migrated from konotor SDK,
                but this is also performed for new installs as well (a harmless side-effect). */
             // TODO : Come up with a better logic to do this migration
             NSNumber *messageLastUpdatedTime = [FDUtilities getLastUpdatedTimeForKey:HOTLINE_DEFAULTS_CONVERSATIONS_LAST_UPDATED_SERVER_TIME];
+            NSMutableDictionary *dictionary = [responseInfo responseAsDictionary][CONTENT_LOCALE];
+            NSNumber *responseLocaleId = [dictionary objectForKey:@"localeId"];
+            if( ![requestlocaleId isEqualToNumber:responseLocaleId] ) {
+                [HLUserDefaults setNumber:responseLocaleId forKey:HOTLINE_DEFAULTS_CONV_LOCALEID];
+            }
+            [FDLocaleUtil updateLocale];
+            
             BOOL isRestore = [messageLastUpdatedTime isEqualToNumber:@0];
             if (isRestore) {
                 [[KonotorDataManager sharedInstance]deleteAllMessages:^(NSError *error) {
-                    [self importChannels:[responseInfo responseAsArray] handler:handler];
+                    [self importChannels:[responseInfo responseAsDictionary] handler:handler];
                 }];
             }else{
-                [self importChannels:[responseInfo responseAsArray] handler:handler];
+                [self hideAllChannelsWithCompletion:^(NSError *error) {
+                    [self importChannels:[responseInfo responseAsDictionary] handler:handler];
+                }];
             }
+        }else if(statusCode == 304){
+            if (handler) handler(nil, error);
+            FDLog(@"No change in channel  data")
         }else{
             if (handler) handler(nil, error);
             FDLog(@"channel fetch failed :%@ \n response : %@",error, responseInfo.response);
@@ -336,7 +354,27 @@ static HLNotificationHandler *handleUpdateNotification;
     return task;
 }
 
-+(void)importChannels:(NSArray *)channels handler:(void (^)(NSArray *channels, NSError *error))handler;{
++(void)hideAllChannelsWithCompletion:(void(^)(NSError *error))completion{
+    NSManagedObjectContext *context = [KonotorDataManager sharedInstance].mainObjectContext;
+    [context performBlock:^{
+        NSManagedObjectContext *ctx = [KonotorDataManager sharedInstance].mainObjectContext;
+        NSFetchRequest *fetchRequest = [NSFetchRequest fetchRequestWithEntityName:HOTLINE_CHANNEL_ENTITY];
+        fetchRequest.predicate       = [NSPredicate predicateWithFormat:@"isHidden == NO"];
+        NSArray *matches             = [ctx executeFetchRequest:fetchRequest error:nil];
+        if(matches.count > 0){
+            for (HLChannel *channel in matches) {
+                [channel setValue:@(1) forKey:@"isHidden"];
+                [ctx save:nil];
+            }
+        }
+        if(completion){
+            completion(nil);
+        }
+    }];
+}
+
++(void)importChannels:(NSDictionary *)channelsInfo handler:(void (^)(NSArray *channels, NSError *error))handler;{
+    NSArray *channels = channelsInfo[@"channels"];
     NSMutableArray *channelList = [NSMutableArray new];
     NSManagedObjectContext *context = [KonotorDataManager sharedInstance].mainObjectContext;
     [context performBlock:^{
@@ -344,18 +382,10 @@ static HLNotificationHandler *handleUpdateNotification;
         NSInteger channelCount = [channels count];
         HLChannel *channel = nil;
         if (channelCount!=0) {
-            for(int i=0; i<channels.count; i++){
+            for(int i=0; i< channels.count; i++){
                 NSDictionary *channelInfo = channels[i];
                 channel = [HLChannel getWithID:channelInfo[@"channelId"] inContext:context];
                 [HLTags removeTagsForTaggableId:channelInfo[@"channelId"] andType:[NSNumber numberWithInt: HLTagTypeChannel] inContext:context];
-                NSArray *tags = channelInfo[@"tags"];
-                if(tags.count >0){
-                    if(!([channelInfo[@"hidden"] boolValue])){
-                        for(NSString *tagName in tags){
-                            [HLTags createTagWithInfo:[HLTags createDictWithTagName:tagName type:[NSNumber numberWithInt: HLTagTypeChannel] andIdvalue:channelInfo[@"channelId"]] inContext:context];
-                        }
-                    }
-                }
                 
                 if (channel) {
                     [HLChannel updateChannel:channel withInfo:channelInfo];
@@ -369,12 +399,19 @@ static HLNotificationHandler *handleUpdateNotification;
                     [channelList addObject:channel];
                 }
                 
+                NSArray *tags = channelInfo[@"tags"];
+                if(tags.count >0){
+                        for(NSString *tagName in tags){
+                            [HLTags createTagWithInfo:[HLTags createDictWithTagName:tagName type:[NSNumber numberWithInt: HLTagTypeChannel] andIdvalue:channelInfo[@"channelId"]] inContext:context];
+                        }
+                }
+                
                 if(channelInfo[@"updated"]){
                     lastUpdatedTime = [FDDateUtil maxDateOfNumber:lastUpdatedTime andStr:channelInfo[@"updated"]];
                 }
             }
         }
-        [[FDSecureStore sharedInstance] setObject:lastUpdatedTime forKey:HOTLINE_DEFAULTS_CHANNELS_LAST_UPDATED_SERVER_TIME];
+        [[FDSecureStore sharedInstance] setObject: channelsInfo[LAST_MODIFIED_AT] forKey:HOTLINE_DEFAULTS_CHANNELS_LAST_UPDATED_SERVER_TIME];
         [context save:nil];
         if (handler) handler(channelList,nil);
         if(channelCount > 0) {
